Move common decoders and encoders to dwio::common
