+++ velox/dwio/dwrf/reader/StripeDictionaryCache.h
@@ -19,8 +19,8 @@
 #include <folly/Function.h>
 
 #include "velox/common/base/GTestMacros.h"
+#include "velox/dwio/common/IntDecoder.h"
 #include "velox/dwio/dwrf/common/Common.h"
-#include "velox/dwio/dwrf/common/IntDecoder.h"
 #include "velox/vector/BaseVector.h"
 
 namespace facebook::velox::dwrf {
