+++ velox/dwio/dwrf/test/TestRle.cpp
@@ -17,8 +17,8 @@
 #include <gtest/gtest.h>
 
 #include "velox/common/base/Nulls.h"
+#include "velox/dwio/common/IntDecoder.h"
 #include "velox/dwio/common/SeekableInputStream.h"
-#include "velox/dwio/dwrf/common/IntDecoder.h"
 #include "velox/dwio/dwrf/test/OrcTest.h"
 
 using namespace facebook::velox;
